Java: consistent oneway behavior
When an interface is marked oneway and a method is called in-process
from Java -> C++, the AIDL compiler still attempts to write out
the status to the reply parcel. This is because interface and methods
have separate oneway values. Here, I've simplified things so that
methods are marked as oneway if they are part of a oneway interface.

This also simplifies API Check (which will be fixed in a separate change)
because the following will be equivalent:
oneway interface IFoo { void foo(); }
interface IFoo { oneway void foo(); }

Bug: 110842805
Test: runtests.sh
Test: android.binder.cts

Change-Id: I7ae7c34383794c97e6e73ba062b2418ea407eb16