Allow 'building' parcelable files.
Before, in Android, if an unstructured parcelable (a.k.a. parcelable
declaration) is present, and it is specified as a source, the build
would fail because:
- AIDL interfaces don't have to start with an I
- the build system must have commands to generate files in its build
  graph before AIDL can be invoked to analyze the file
- the build system would assume that a file (e.g. Foo.aidl) is an
  interface and would generate output
- AIDL didn't generate output for parcelables, so if Foo.aidl was
  solely just a parcelable, then the build system would error.

Now for parcelables, we generate a mostly empty java/cpp file (comment
with small explanation) for the expected output.

In a follow-up CL, I will completely remove the 'FailOnParcelable'
option since it provides no value.

Bug: 70046217
Test: runtests.sh

Change-Id: I5d33b5b8b4ec449dd94b3069ecbb34401603f8be