aidl_interface: support filegroups
Before we were processing files at load hook time. This was because for
a CreateModule, errors were attributed to the created module instead of
the root module. Now that this creation information is tracked and
displayed by the build system. So, we can process source files later
when we have dependency information required to lookup filegroups.

Bug: 133112046
Test: build test interfaces
Test: manually try adding non-aidl file, having no sources
Change-Id: Icaba28356152780052922e3fcab4478f1608fa24