Match src paths with aidl package name
In order for the build system to track updates to the header files
during incremental builds, always specify the src files using the same
path as the package for C++ compilations.

Also writes out the three interface headers from parcelables so that we
can always expect the same output files from all *.aidl files. This is
the same workaround used by StructuredParcelables.

Bug: 112114177
Bug: 111362593
Test: treehugger
Change-Id: I0b78f92fa0e77c67c3a92d8e281f0905336dd050