Allow multiple declarations of unstructured parcelables in a file
AIDL compiler now does not complain when there are multiple declarations
of unstructured parcelables in a file. This is needed to pass AIDL files
under frameworks/base captured with glob pattern (e.g. **/*.aidl). Some
of the AIDL files in the directory are for unstructured parcelables and
they sometimes have multiple declarations in a single file. This is not
bad since we don't generate any code for unstructured parcelables and
thus should not complain about this.

Bug: 70046217
Test: m
Change-Id: Iad35f46452858b99fca51c80cb499179f4a8503c