Check stable AIDL interfaces based on hash
A dump of an AIDL interface now includes hash of the all AIDL files plus
the version number. The hash is recorded to .hash file under the API
directory and is compared when comparing ToT with the latest frozen
version of the API.

Bug: 136065010
Test: cherry-pick the CL to qt-dev where PLATFORM_VERSION_CODENAME is REL,
add random values to the .hash files in the API directories. Build breaks.

Change-Id: I1d49c9d6e6bcc6bd7bf16f0a524e61a52263cea5