Disable Java backend for VINTF.
It's not clear yet how we will support stability checks in Java. For
now, making sure that the Java backend is disabled when vintf stability
is enabled.

Some things we are looking at:
- moving image mutator up (big build change)
- having aidl_interface/aidl: build system inform `aidl` whether to mark
  system or vendor stability.

Will re-enable this once there is a solution in place for Java, but C++
HALs should still be possible.

Bug: 139325468
Test: TH
Change-Id: Ifd9a8b058a5854b22b7dd467424f6cda11ea1e9b