Include source position in generated annotations.
There is a bunch of tooling build on top the annotation index created for
the UnsupportedAppUsage annotations. That index currently points to the
java code generated from AIDL. Including the original source position here
allows the index to point the the original location in AIDL, simplifying
the downstream tooling.

To get the exact source location needed, modify the location generated in
aidl_language.yy to span the whole annotation including the parameters.
The location end now corresponds to the closing parenthesis if there are
any parameters on the annotation.

Bug: 145120552
Test: m
Change-Id: Ie6387394bf6403766000ba0b3654ca47619ab23d