aidl: sanitize for integer overflows
Hidden overflows are extra painful in AIDL interfaces because they may
be stabilized. In order to make sure values in interfaces are exactly as
expected, we are turning on hard failures here.

This should make the fuzzer more effective (discovering these edge
cases). In the mean time, this will prevent intefaces which have
undefined overflows in constant expressions from getting created.

Future considerations:
- remove duplication of tests between constant expressions as constant
  variables and in lists, where they are easier to test.
- better compiler errors for overflows (once the fuzzer catches them)

Bug: 148149098
Test: TEST_MAPPING

Change-Id: I4f936b96dc6eb7714ae622a444ebcc64da17a934