AIDL: -freeze-api includes vintf stability
aidlApi now has a stability parameter and properly passes stability.

We may consider combining aidlApi and aidlGen: 'dumpapi' is very similar
to what "--lang=dump" would be, and we could avoid a lot of duplicated
logic (doing something similar for checkapi could potentially remove
some of the compilation differences we have for checkapi now as well).

Bug: N/A
Test: m android.hardware.vibrator-freeze-api
Change-Id: I3542c6aec5a36a3a83b917ef8bae000c188dac64