Limit VNDK entries to latest frozen version.
Given 'aidl_interface { name: "foo" }'
we generate (a) 'foo-V1-ndk_platform.so'

Given 'aidl_interface { name: "foo", versions: ["1"] }'
we generate
(b) 'foo-V1-ndk_platform.so'
(c) 'foo-V2-ndk_platform.so'

We want the latest version to be in the VNDK. So (a) and (c) get put in
the VNDK. In the (a) case, the VNDK list will be updated during the
release. In the (c) case, when version 2 is added by freezing this
release, the entry in the VNDK list will be updated.

This is to reduce the total number of libraries in the VNDK (since
having (b) and (c) would be too much). Ultimately, we might remove all
of these stubs from the VNDK. However, for now, this is a hedge against
the AIDL hashing infrastructure.

Bug: 149960746
Test: freezing android.hardware.vibrator (which has no versions) allows
    the build to succeed without having to modify the VNDK current.txt
    file.

Change-Id: Ifbeafbbf86a8908c9be7a29e39889eed7c5a6213