aidl_integration_test: do all supported combos
Before, this tested a Java or first-bitness native client with a
first-bitness native server.

Now, according to the supported bitnesses on device, this runs a Java
client and native clients of supported bitnesses against native servers
with supported bitnesses.

Note: sync and run works for native tests, but not Java, since the APK
needs to be installed.

Bug: 129507417
Test: atest aidl_integration_test (on walleye and cuttlefish)
Test: introduce failure in native and java test and see that they fail
Change-Id: I3c869fdf6758217b7c3fc3b42d525bdf7cbb7b48