Changing enum autofill to use aidl file location instead of cpp
This location is printed when there are any issues with the aidl node.
In this case, when implicitly starting enumerators from 0 or implicitly
adding 1 to the next enumerator, we are now using the location of that
enumerator in the .aidl files instead of the .cpp file. This is intended
to give more useful information to the user.

Added a check to make sure we don't output internal file locations.

Test: atest aidl_unittests aidl_integration_test
Change-Id: I0b2b7a18e69c313668aea395a2a02a77ae27dfc1