DESCRIPTOR is composed at runtime for stable AIDL interfaces
When java stub classes for an AIDL interface is jarjar'ed in either of
the client or the server side, the peers can't talk with each other.
This is because the client and the server checks whether they are using
the same interface name. The client sends its own descriptor name, and
the server checks if it matches with its own descriptor name. The
descriptor name is stored in the DESCRIPTOR field (static final) of the
interface class. When the interface class is jarjar'ed, the string
literal in the field is also renamed, thus invalidating the check.

The idea here is to compose the descriptor name at runtime so that it is
not renamed by jarjar. Below is how the descriptor was defined.

private static final String DESCRIPTOR = "android.foo.IMyInterface";

This CL changes above into the following.

private static final String DESCRIPTOR = "android$foo$IMyInterface".replace('$', '.');

Since the string literal android$foo$IMyInterface does not match with
android.foo.IMyinterface, it is not renamed by jarjar.

Note that this change is only for stable AIDLs (i.e. --structured).

Bug: 153843174
Test: run jarjar on an aidl interface (netd_aidl_interface for example)
and inspect the output to see if the string literals with $ are not
altered.
Change-Id: If4c0ecafd855d108c8067869d29651004a13655c