Add support for atrace in new build system
Support was added for the old build systems, but was never added for
Soong aidl_interface.
Tracing in AIDL interfaces is now enabled by the gen_trace
property in the aidl_interface blueprint.

Test: atest aidl_unittests aidl_integration_test
Test: enabled tracing in netd/server by adding 'gen_trace: true'
to the instances of aidl_interface in netd/server/Android.bp. Built with
'm netd'. Verified the presence of the atrace_begin and
atrace_end function calls in the generated source at
$OUT/soong/.intermediates/system/netd/server/.
Bug: 150952603

Merged-In: I76d15b4dbcd21cdf2200b3e997477a096c5469a9
Change-Id: I76d15b4dbcd21cdf2200b3e997477a096c5469a9