Annotation: expose annotation type with enum class
This is in preparation for 'checkapi' allowing/disallowing certain
annotations being added/deleted. Before, it's impossible to see what
type of annotation a specific AidlAnnotation. APIs like 'IsHide' etc..
only work on AidlAnnotable, but can not be used on individual
annotations.

This ensures that when we talk about an annotation, that annotation
is guaranteed to exist (e.g. can't typo "nillable" for instance).

Future considerations:
  - We could remove 'GetName' of AidlAnnotation
  - IsHide/etc.. can be replaced with generic API

Bug: 146215188
Test: atest --test-mapping .
Change-Id: If45942385cb92c32d5ae9ba8e2a3bd9f044ecbb6