Add Atrace support for aidl_interface NDK backend
This is enabled the same way that it is for the Java and C++ backends
with gen_trace option in the aidl_interface defintion. Traces can be
captured with atrace on device.

Test: Added 'gen_trace: true' to libbinder_ndk_test_interface and
collected traces from CtsNdkBinderTestCases with perfetto.
adb shell perfetto -d -o /data/misc/perfetto-traces/out.trace --app android.binder.cts

Bug: 128533039
Change-Id: Id14b391e59151adc51688d5f295b8627bff56d48