Untyped List or Map is prohibited for new interfaces
The use of Untyped List or Map is prohibited in the stable AIDL
because it is not guaranteed that the objects in the container are
recognizable in the receiving side (or are in sync across the sender and
the receiver).

This change rejects any addition of a method or a field using such
untyped List/Map during the API check. The existing uses are allowed
for backward compatibility reason.

Bug: 155399466
Test: aidl_unittests
Change-Id: Ia190b9883d6dbc0f8f06c32165025aaa9cce0b14