Remove AidlQualifiedName
AidlQualifiedName is removed from AST in favor of std::string. It looks
like the class was introduced to handle the language-specific handling
of the qualified names (e.g. foo.bar becomes foo::bar in C++).
Since such language-specific routines are all moved to the backend,
we don't need to have the routine in the frontend.

Bug: 156309715
Test: m
Test: run aidl_unittests

Change-Id: I8acf43d804e92e79d8a588cdb429fdc68d625cbf