Fix a memory leak
Consider this case: parcelable m{}t

`parcelable m{}` is parsed into a AidlStructuredParcelable object and
the pointer to it is put to a vector. Then the parsing error occurs
due to the non-parsable `t`. As a result, the vector is destroyed.
But since the AidlStructuredParcelable is stored as a pointer, it's
not destroyed - leaked.

Fixing this issue by defining a custom destructor for the vector so that
its elements are also destroyed along with the vector.

Bug: 159238734
Test: $ SANITIZE_HOST=address m aidl_parser_fuzzer
$ cat " parcelable m{}t" > input.aidl
$ out/host/linux-x86/fuzz/x86_64/aidl_parser_fuzzer/aidl_parser_fuzzer
input.aidl
Change-Id: I7f51e0801db1283762cd0580605b2e14a2e24d69