Don't use DISALLOW_COPY_AND_ASSIGN macro
Instead of using the macro, explicitly delete the copy constructor and
copy assignment operator. In addition, move constructor and move
assignment operator are now deleted when they are not needed. The
explicit deletion of the move is actually not necessary because they are
not auto-generated when the copy constructor/assignment operators are
explicitly declared. But they are deleted to prevent them from being
declared in the future.

In ast_cpp.h, only the AstNode class is marked as non-copyable and
non-movable. The child classes don't need to delete the copy/move
functions because they can neither be auto-generated nor default defined
because they both require copy/move function in AstNode, where they are
deleted.

Bug: 161855722
Test: m
Change-Id: Id502cf4b62f3d1d513a72d4dd21fd83f2af8ad9c