Add tests for getInterfaceHash and getInterfaceVersion
aidl_integration_test is added with a few more tests for
getInterfaceHash and getInterfaceVersion. For this, a new versioned AIDL
interface "aidl-test-versioned-interface" is newly defined. It has two
versions 1 and 2. The clients (written in Java and C++) use the newer
version and the service (written in C++) use the old version. Tests are
added to the clients to ensure that the version and the hash are those
of the version 1, not the version 2.

Bug: 150164755
Test: atest aidl_integration_test
Change-Id: I96b80a4a6aef903857478d3c2a54c69cbb997f77