Support @FixedSized annotation for structured parcelables
This annotation marks the parcelable as fixed size so it can be
guaranteed to stay the same size.
This can be particularly useful for passing these types over shared
memory between processes that are updated independently. No matter which
version either process is using, the type will have the same size.

Test: atest aidl_unittests
Bug: 142326204

Change-Id: I18240e2832da8578bee077026955686ec90c0e5f