Change the rule to pick up VNDK targeted aidl interface version
For a non-versioned module,
As-is: a module without postfix
To-be: a module with '-unstable'

The contents of them are identical, just prevent duplication in VNDK lib
list due to imports.

Todo: In the following CL, enforce VNDK target module not to be
'unstable'. To prevent unstable module to be included in VNDK.

Test: m
Bug: 146436251
Change-Id: Iea2618900fbce4a0f45f8fdc4ff1a5cf530d87de