Descriptor annotation is added
The descriptor annotation allows the author of an AIDL interface to
override the interface descriptor token which by default is set to the
fully qualified name of the interface. For example,

package android.foo;

@Descriptor(value="android.bar.IWorld")
interface IHello {...}

The descriptor of the above interface is "android.bar.IWorld". If the
Descriptor annotation was missing, it would have been
"android.foo.IHello".

This would be useful when an already published interface has to be
renamed. By making the descriptor of the renamed interface to be the
descriptor of the interface before the renaiming, the interfaces -
before and after the renaming - can talk to each other.

Bug: 165712106
Test: atest aidl_unittests
Test: atest aidl_integration_test
Change-Id: I9cee8d883925f4d53adde4a131af450fb689b269