RuntimeException on unknown transaction
Previously, a call to a non-existing method was silently ignored when
the default implementation was not set for the interface. Now, it
triggers a RuntimeException in the client side. Note that the exception
is thrown only when the interface is versioned. For unversioned
interfaces which have been available to apps, this would be a non-compat
change.

Bug: 167632276
Test: aidl_unittest, aidl_integration_test
Change-Id: I58ac249e85a67a2b0671d66ef208eeb81588641d