aidl_interface: avoid vendor_available contagion
Before, for aidl_interface modules A and B, if A imports B and B has
vendor_available: true, then A would also need to set vendor_available.
However, this restriction isn't a desirable restriction and the error
message ("A-ndk depends on undefined module B-unstable-ndk") wasn't very
clear.

This restriction was added because ndk (here, meaning library built to
run inside of an application against the ndk) libraries interact poorly
with image variant properties like vendor_available. The build system
won't allow you to specify the "vendor" version of something in an app
because this doesn't make sense. So, in this situation, before, we
didn't create the NDK variant in B. However, A's NDK variant would still
depend on this, because A->B.

In order to simulate the NDK being its own orthogonal image variant,
here we disable image variants with a bad interaction (vendor_available,
vndk) explicitly. So, now, we still generate the app-specific library,
but we manually disable all of the properties we know interact badly
with it so that we have what looks like a library specifically designed
for apps.

Note, although aidl_interface supports building interfaces for apps, for
a specific interface, it's expected to be entirely used within the
platform or entirely used within apps. However, for simplicity, we
create all of the variants that may be used.

Bug: N/A
Test: simplified build tests
Change-Id: I85e940e1d692adb7a958a898ff1bedde0ef5c83a