Remove FixedSize handling from cpp backend
The @FixedSize annotation is to be used for guarenteeing the size of the
type to allways be the same.
In the cpp backend parcelables inherit from an abstract Parcelable class.
The vtable pointer is different sizes when built for 32-bit or 64-bit
which can cause the size of the final parcelable to be different.
This also means that these objects can not be passed to another process
over shared memory, which is the current use case for @FixedSize.
The ndk backend does not have this issue, so @FixedSize will still be
supported there.

Test: atest aidl_intergration_test aidl_unittests
Bug: 142326204
Change-Id: I389eb8ff377e47ced8da06b12e9f301fbc0aa8cc