ndk: Emit aligment values for known types in @FixedSize parcelables
This is only done in the ndk backend. It ensures that the @FixedSize
parcelable is the same size accross different builds.
For example:
uint64_t has an alignment of 4 in 32-bit bit builds, 8 in 64-bit, making
it possible for it to be a different location on some devices.
Using __attribute__((aligned(8))) in all cases ensures the same layout.

Test: atest CtsNdkBinderTestCases aidl_integration_test aidl_unittests
Test: built CtsNdkBinderTestCases for aosp_x86_64-eng before/after this change
Bug: 142326204
Change-Id: I0be49db94e911df37acfc2b0bd3578715975a3d1