integration tests: avoid setenforce
setenforce disables SELinux globally on a device. This was required
because even as root, requesting an application to be run does not give
that app root priveleges. So, we could not retrieve services. However,
by globally disabling SELinux, we potentially disrupt the system as well
as potentially rely on behaviors that are disallowed (when this code
talks to other processes).

So, this:
- moves the test from an app into a jar
- uses 'app_process' to launch our desired functionality as root

Bug: 169704480
Test: atest aidl_integration_test
Change-Id: Iad9a1bc98809867673c2560f7b23f36dca098de0