Show full dependency path
checkDuplicatedVersions now shows the full dependency path when it finds
an error. This is done by separating it into two mutators. The first one
gathers (direct and indirect) dependencies to the AIDL interfaces, and
the second one checks if there is any duplicated versions. If there is,
the mutator runs WalkDeps until the aidl_interface (actually the
generated java, c++ modules) is hit and prints the dependency path using
GetWalkPath.

Bug: 170427726
Test: m
Change-Id: Ifae60aaba5f33208566977186e4565ca873e2c8c