Fix Rust AIDL handling of multiple imports
When generating the Rust code for an interface with multiple imports,
we only generated one "pub(crate) use" statement.  This seems to be
because the Python parser that parsed the arguments expected include
arguments of the form "-I a b c" but instead received "-Ia -Ib -Ic".
This modifies the parser to expect arguments of the second form.

Test: Compile an interface with multiple includes and see both in the generated source.
Change-Id: Id8b2a5bc814c9c04895f4b1a25d0d0376bc8d9b0