Update API preamble (talk about 'current' case)
The API preamble goes on API files, but on 'current' files, this was
confusing some people, because it seemed like these files could not be
changed.

Here, I went the easier approach of explaining both types of API files
in the header, rather than generating a different header for 'current'
API files. This may be confusing, but on the other hand, it avoids
additional arguments/build complexity, and the additional information is
probably helpful even on frozen API files - sure those files can't be
changed, but this explains how later 'current' API snapshots can be
updated.

Fixes: 175801423
Test: update all 'current' APIs in this directory, with this
  very hacky bash one liner (and confirm things build):

    for i in $(grep -rn "aidl_interface {" -C2 --include="Android.bp" \
        | grep name: | grep -oP "(?<= \")[^\"]*" \
        | sed s,$,-update-api,); do m $i; done

Change-Id: I90cf76f526800f7c65da063aecc0ea4c32fdf2eb