AidlAnnotation::ParamValue() to retrieve param's value
Similar to AnnotationParams() which returns a map<string, string> with
all params and their stringified values. Even though this is useful when
we pass them to generated Java code, it's not when the AIDL compiler
needs to process annotations.

ParamValue(param_name) looks up the parameter and returns the typed
value.

Bug: none
Test: aidl_unittests
Change-Id: I99aa8e6c0fd787350b02d3e84b642e387d41a40e