Refactor AidlVisitor so that it no longer responsible for traversing
Visit() methods now return a bool which controls whether or not the
traversal should be continued down to its children or not.

A new trait AidlTraversable is introduced. The type is expected to be
implemented by an AST node which needs to be traversable. The following
three methods need to be implemented:

* TraverseChildren: calls the given traverse function on each of the
traversable children, if there is any.
* DispatchVisit: calls the corresponding Visit method

Traversing the AST is done by any functor that accepts AidlTraversable.
Inside the functor, the various traversing strategies (Top-Down,
Bottom-up, etc.) and stateful options (pushing/popping the diagnostics
options which will be the follow-up change) can be implemented.

Bug: N/A
Test: aidl_unittests
Change-Id: I27e67f53429641b6cdb5727929b1bff77d4d4163