Emit deprecation note
@deprecated tag in comments can have optional note to explain the
rationale for deprecation and/or to suggest a replacing entity.

  /** @deprecated use bar() */
  void foo();

In this change, the deprecation note is emitted when specified.
For example, in C++, deprecation note is emitted as following:

  void foo() __attribute__((deprecated("use bar()")) ...

Bug: 174514415
Test: aidl_unittests
Merged-In: I271af4d636e755a4d97f60d2daa6528acbe76a4e
Change-Id: I271af4d636e755a4d97f60d2daa6528acbe76a4e
(cherry picked from commit 6ec07c5482939803229445726044457383dbdaa9)