Lexer passes the list of comments to parser
The lexer has been squashing comments into a single string, which makes
it hard to distinguish the type of comments later.

Instead of re-parsing comments, now, the lexer keeps the list of
comments and pass it to the parser.

But note that we still need to parse the contents of the body of block
comments to extract @hide/@deprecated tags.

Bug: 177276676
Test: aidl_unittests
Change-Id: I5f683ba6bf9fa3fbc5e73c4f80237a25eb410f71