refactor: AidlAnnotation/AidlTypeSpecifier are commentable
These nodes are not syntactically commentable, but they hold comments
for parent nodes while parsing.

For example,

  // comments
  @Annotation int foo;

The comment is first attached to the annotation and then transfered to
the type('int') and then finally passed to the variable declaration.

While processing AST, we only care about variable 'foo's comments.

Bug: none
Test: aidl_unittests
Change-Id: I11f19dd9e5f8b5fe6a8343f862bbe34393ad8a4d