aidl-format: .aidl-format for further customizing
When the default style is not desirable (e.g. 2-space indent rather than
4-space indent), .aidl-format file can be used to further customize the
default config. The format of .aidl-format is the same as .clang-format.
When an AIDL file is being formatted, .aidl-format file is found in the
nearest parent directory of the input file. If stdin is the input, it's
found from the current directory.

Bug: 144540481
Test: cat > .aidl-format
IndentWidth: 2,
^D
./aidl-format.sh tests/android/aidl/tests/ITestService.aidl

The result is with 2-space indent

Change-Id: Id833ea9fd2fdf685a1aecc00bc2753c96598996a