run_integration_test - work again
Instead of installing an apk, we now copy a library to the /data/
partition. This change was missed when I removed the app, to avoid
needing setenforce 0 inside of the test.

Still keeping/maintaining this test, because it's simpler/easier/faster
than running atest (and we still need atest to plug into
infrastructure).

Bug: N/A
Test: ./run_integration_test.sh
Change-Id: Ibb0eebc2fec19d5a045f9979bcfe2adf083e2455