syntax: annotations before [] and type args
This is a re-write of the grammar to support "T @annotation []" or
"List<@annotation T>". (To be specific, TYPE_USE annotations in Java).

However, the compiler doesn't support those annotations yet and rejects
input with TYPE_USE annotations.

This will make it easy to support them when we want(or are ready) to.

Bug: 174862286
Test: aidl_unittests
Change-Id: Idf0a0aeda6fd33cbdd225276cb9a5ea36f29f59b