fix: handling Map's key/value type
Map's key type should be "String". But "String[]" was treated OK and the
generated Java code was wrong. (Failed to compile).

Key type should be check via "Signature()" instead of "GetName()".

And, createFromParcel() for value type was wrong when the value type is
array. This is fixed by using JavaSignatureOf() instead of JavaNameOf().
(The former handles "array" while the latter doens't.)

Bug: 180255700
Test: aidl_integration_test
Change-Id: I00c548f469952ba901df55e6bba486e82eeee810