AIDL test Java: getService -> waitForService
getService is racing with registration of services, especially for Java
clients, since in the Java backend, getService does not wait 5s for
services to appear (like it does for other backends).

This means, if it fails, we'll get timeout instead of a failure. We
could use a constant timeout here, but I've never had a problem with
waiting indefinitely in HIDL tests - it's flake proof! If there is a
server crash causing a timeout, it should be easy to repro locally.

Fixes: 179698026
Test: (runs for a long time w/o repro)
    while ./tests/aidl_integration_test.py; do adb logcat -c; done
Change-Id: I04ee9e87806fa4c5a3f90083e15a6d3c7cb4998d