More friendly error when invoked from repohook
When multiple errors are found, repohook doesn't provide the automatic
fixup. This however can be confusing to the users because they might not
know how to fix the errors. This change provides a simple instruction
about how an AIDL formatting error can be fixed.

Note that we can't show the input file path because repohook passes it
via stdin.

Bug: 181078995
Test: apply ag/13680608, run `repo upload`.
[FAILED] aidl_format
  FILES: ('biometrics/face/aidl/android/hardware/biometrics/face/SensorProps.aidl',)    85d84
    <
    You can try to fix this by running:
    .......system/tools/aidl/aidl-format.sh -w <file>
      FILES: ('biometrics/fingerprint/aidl/android/hardware/biometrics/fingerprint/SensorProps.aidl',)    62d61
    <
    You can try to fix this by running:
    .......system/tools/aidl/aidl-format.sh -w <file>

Change-Id: I46388dc14911db5257afac37e0a0b4e848edb08b