Versioned test to verify readFromParcel
The added tests veify that parcelable's readFromParcel correctly skips
extra data from newer version so that the subsequent reads work
correctly.

For example, in C++, if we miss the setDataPosition at the end of
readFromParcel(), the readInt32 for the second parameter (value) reads
the wrong value from the unread extra data of the parcelable.

Bug: 186633641
Test: aidl_integration_test
Change-Id: I6b16bc363b3725f934c735355911cc03e35a612a