Reland: NDK backends supports both apps and platform
Previously, the ndk backend was for apps, and the ndk_platform backend
was for platform. However, this has not only caused confusion to the
users, but also added unnecessary complexity as the ndk backend actually
has "two" variants 'sdk' and 'platform' by the sdkMutator in Soong. The
'platform' variant can replace the ndk_platform backend.

This change is the first step to remove the ndk_platform backend. The
'platform' variant of the ndk backend now has the same dependencies as
the ndk_platform backend. This is achieved by using the target specific
properties which adds and removes dependencies for the non-app variants.

The next step will be actually replacing the references to
*-ndk_platform modules with the corresponding *-ndk modules. After that
is done, the ndk_platform backend will be removed.

Ignore-AOSP-First: needs c/p's

Bug: 161456198
Test: m
Change-Id: Id5f5a4b218b601907c306bc8287c0d511ad84668