aidl_parser_fuzzer: fuzz everything
Fuzz the CLI - we're not playing games anymore.

Leaving corpus for now, because the contents at least are known to be
useful to fuzz, even though the fuzzer requires a new format now. We
could remove them later.

Bug: 195473218
Test: aidl_parser_fuzzer locally, and it was able to make some progress,
for instance, I can see we get here, we know we at least shouldn't be
regressing on coverage.
    ERROR: :1.1-3: 'j' is not a recognized annotation. It must be one
    of: nullable utf8InCpp SensitiveData VintfStability
    UnsupportedAppUsage JavaOnlyStableParcelable Hide Backing
    JavaPassthrough JavaDerive JavaOnlyImmutable FixedSize Descriptor
    RustDerive SuppressWarnings.

Change-Id: I8b72b8b057ce9dceb178e5b9bf8e34e27d48d2b3