import versions for tests
This makes sure that when developers add a new version of an interface,
or when interfaces are being frozen, the runtime/buildtime situation of
clients depending on those interfaces remains the same. This is required
for AIDL to continue working at scale.

Note: some tests are changed because with this enforcement turned on,
it's no longer possible for the latest version of an interface to depend
on a different version of its dependency from the previous version of an
interface.

Bug: 188871598
Test: build
Change-Id: Ic4c8f6ad05a79d267a991999f56fb6f030d739fb