DO NOT MERGE: Reland: Don't generate the ndk_platform backend unless explicitly requested
We can't completely remove the support for the backend because there
still are devices (outside of AOSP) whose BSP depends on the
ndk_platform backend and we know that those BSPs can hardly be fixed
because they are outsourced. Let's have the switch for the time being,
but the default behavior is chagned to "no ndk_platform backend".

Bug: 161456198
Test: m
Merged-In: Ia3d400236b831c6052278e71949f2afcd53c25ca
Change-Id: I763495eaf1abfaf011c3fcdc3fbe9282d74c19cb