Don't generate the ndk_platform backend unless explicitly requested
We can't completely remove the support for the backend because there
still are devices (outside of AOSP) whose BSP depends on the
ndk_platform backend and we know that those BSPs can hardly be fixed
because they are outsourced. Let's have the switch for the time being,
but the default behavior is chagned to "no ndk_platform backend".

Ignore-AOSP-First: need to check this for internal targets first

Bug: 161456198
Test: m
Change-Id: I1cbdce9162e08f54799b42004ef5326adf288ba4