Add further permission annotations
Introduce the @NoPermissionRequired annotation to explicitly declare
that an interface or method does not require any kind of permission when
called.

Add the @PermissionManuallyEnforced annotation to indicate that
permissions are manually verified within the implementation.

Ensure that these new annotations are not defined in a conflicted manner
with the @Enforce annotation.

Bug: 197828948
Test: atest --host aidl_unittests
Change-Id: Iff3b4c9c5d2a5b4c310d7a737b8d7628cc39ba20