Fix nested-interface in NDK with log/traces
This fixes two problems:
- ScopedTrace class redefinition
- Wrong qualification for logFunc variable

In the NDK backend, gen_traces option requires an extra definition of
ScopedTrace class. Since it's defined in the target namespace root, when
there's an interface-in-interface, the generated code was broken with
duplicate definition.

Now it's defined only once in anonymous namespace in a file.

logFunc variable is now qualified with parent types.

Bug: n/a
Test: m aidl_test_loggable_interface-ndk
Test: aidl_integration_test
Change-Id: I55556248937ab0347cd9259112a92bf341b48dae