Add @FixedSize union
@FixedSize union uses a custom implementation rather than std::variant
so that its value can have a stable/fixed layout and be passed via FMQ.

A custom implementation relies on the fact that @FixedSize
union/parcelable should have "trivially copyable" fields. So we can't
use it for union in general.

It would be nice if we could replace std::variant with it and have the
same implementation for union types. Maybe in the future.

Bug: 206496560
Test: golden_test.sh check
Test: aidl_integration_test (no regression)
      atest fmq_test fmq_unit_test
Change-Id: I86351ecb1990818d0fe2faaad7c4b8c2c33881b5