Use PermissionChecker for Java backend
Move to PermissionChecker.checkPermission to validate the permissions.
Use the ForDataDelivery variant as default (the Preflight variant can be
added later on, using a different predicate for example).

If the method receives an AttributionSource as parameter, it will be
automatically used for the permission check. Otherwise, a new
AttributionSource will be generated.

Ran bpfmt on Android.bp, some lines have been reformatted.

Bug: 197828948
Bug: 208707422
Test: atest --host aidl_unittests
Change-Id: I0ca010eafba0fc02a614f1add9275dd92938d623