Restore character behavior.
This was originally done in this commit:
Revert "Remove string literal checks."

But due to a conflict, landing this in a slightly different way.

The idea is, we don't want to allow arbitrary codepoints in a file,
and we want to make sure we know explicitly what the content of a string
is.

Bug: 215037384
Test: aidl_unittests
Ignore-AOSP-First: related to security fix
Change-Id: Ic16e0b5b860395cda80af1ca9bf3bd9d24d973ab