AIDL: add propagateAllowBlocking
This allows us to specify how allowBlocking is propagated without
having the SDK or SystemApi allow clients to choose which
interfaces are allowBlocking. Although in the current design,
this decision is made by clients, by only exposing this
relationship here, we can allow the system to choose which
interfaces from where can make synchronous calls.

Bug: 216685683
Test: see golden_output
Change-Id: I3385bdf350fea8ef709651773b989bbc7afe671a