Java: require interfaceVersion/hash for ToT vers.
This reqiures source changes, so we need to require it in order to
automate API finalization. The reasons for not having this for the
development version are mostly from the legacy C++ backend. For
minimalism here, the NDK and Rust backends are not considered.

Bug: 226954348
Test: build requires interface version and hash on interfaces.
Merged-In: I0b350d9cbc4aca62bff6dc268ee7419044fc6d01
Change-Id: I0b350d9cbc4aca62bff6dc268ee7419044fc6d01
(cherry picked from commit ed0490c75a339a9c997827282a827fe721836e06)